Current state of unit tests as of 27 Nov 2023
